+++ lib/module/advertising/affise_module_advertising_api.dart
@@ -1,0 +1,5 @@
+import '../affise_has_module.dart';
+
+abstract class AffiseModuleAdvertisingApi extends AffiseHasModule {
+  void startModule();
+}