+++ lib/module/attribution/affise_attribution_module_api.dart
@@ -1,14 +1,13 @@
 import '../../affise.dart';
 
 abstract class AffiseAttributionModulesApi {
+  AffiseModuleAdvertisingApi get advertising;
   AffiseModuleAppsFlyerApi get appsFlyer;
   AffiseModuleLinkApi get link;
   AffiseModuleSubscriptionApi get subscription;
   AffiseModuleTikTokApi get tikTok;
 
   getStatus(AffiseModules module, OnKeyValueCallback callback);
-
-  Future<bool> moduleStart(AffiseModules module);
 
   Future<List<AffiseModules>> getModulesInstalled();