+++ lib/module/export.dart
@@ -8,3 +8,4 @@ export 'link/affise_module_link_api.dart';
 export 'subscription/export.dart';
 export 'appsflyer/affise_module_appsflyer_api.dart';
 export 'tiktok/affise_module_tiktok_api.dart';
+export 'advertising/affise_module_advertising_api.dart';